# openapi.yml
openapi: 3.1.1
info:
  title: Fitness tracker API
  description: API for tracking fitness activities
  version: 1.0.0
paths:
  /exercises:
    get:
      summary: Get all exercises
      description: Retrieves a list of all exercises available
      responses:
        200:
          description: A list of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
  /workouts:
    get:
      summary: Get all workouts
      description: Retrieves a list of all workouts for the user
      responses:
        200:
          description: A list of workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
  /workouts/{uuid}:
    get:
      summary: Get workout by UUID
      description: Retrieves a specific workout by its UUID
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
          description: UUID of the workout to retrieve
      responses:
        200:
          description: A workout object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
        404:
          description: Workout not found
  /workouts/create:
    post:
      summary: Create workout
      description: Creates a new workout for the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkout"
      responses:
        201:
          description: Workout created successfully
        400:
          description: Bad request, invalid input data
components:
  schemas:
    Exercise:
      type: object
      properties:
        uuid:
          type: string
          description: UUID of the exercise
        name:
          type: string
          description: Name of the exercise
        description:
          type: string
          description: Description of the exercise
        restTime:
          type: integer
          description: Rest time in seconds between sets
    WorkoutExercise:
      type: object
      allOf:
        - $ref: "#/components/schemas/Exercise"
        - type: object
          properties:
            sets:
              type: array
              description: List of sets for the exercise
              minLength: 1
              items:
                $ref: "#/components/schemas/ExerciseSet"
    ExerciseSet:
      type: object
      properties:
        reps:
          type: integer
          description: Number of repetitions for the set
        weight:
          type: number
          format: float
          description: Weight used for the set
        duration:
          type: integer
          description: Duration of the set in seconds
    Workout:
      type: object
      properties:
        uuid:
          type: string
          description: UUID of the workout
        name:
          type: string
          description: Name of the workout
        exercises:
          type: array
          description: List of exercises in the workout
          items:
            $ref: "#/components/schemas/WorkoutExercise"
    CreateWorkout:
      type: object
      properties:
        name:
          type: string
          description: Name of the workout
        exercises:
          type: array
          description: List of exercises in the workout
          minLength: 1
          items:
            $ref: "#/components/schemas/CreateWorkoutExercise"
      required:
        - name
        - exercises
    CreateWorkoutExercise:
      type: object
      properties:
        uuid:
          type: string
          description: UUID of the exercise
        sets:
          type: array
          description: List of sets for the exercise
          minLength: 1
          items:
            $ref: "#/components/schemas/ExerciseSet"
      required:
        - uuid
        - sets